requirements.txt
pydantic>=2.6
jsonschema==4.22.0
python-dotenv>=1.0.1
requests>=2.31.0
streamlit>=1.35.0
jinja2>=3.1.4
